% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-extra-checks-pkg.R
\docType{package}
\name{pkgcheckEditorExtra-package}
\alias{pkgcheckEditorExtra}
\alias{pkgcheckEditorExtra-package}
\title{pkgcheckEditorExtra: Extra rOpenSci Package Checks for Editors}
\description{
The main 'pkgcheck' package checks whether a package is ready for submission to rOpenSci's peer review system. This package contains additional checks for our editors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgcheck-editor-extra/}
  \item \url{https://github.com/ropensci-review-tools/pkgcheck-editor-extra}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgcheck-editor-extra/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
